Refactored merge function
